Avoid overflow in cuesheet marker positioning. Fixes FS#12246. Thanks to Nick Peskett.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30359 a1c6a512-1295-4272-9138-f99709370657

+++ apps/cuesheet.c
@@ -364,7 +364,7 @@ void cue_draw_markers(struct screen *screen, struct cuesheet *cue,
     int i,xi;
     for (i=1; i < cue->track_count; i++)
     {
-        xi = x + (w * cue->tracks[i].offset)/tracklen;
+        xi = x + (w * (long long)cue->tracks[i].offset)/tracklen;
         draw_veritcal_line_mark(screen, xi, y, h);
     }
 }
